requirements.txt
scrapy==2.13.1
openai==1.84.0
python-dotenv==1.0.0
chromadb
langchain
langchain-community
langchain-openai
langchain-chroma
langchainhub
fastapi
uvicorn